requirements.txt
appnope==0.1.0
backcall==0.1.0
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-Cors==3.0.8
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
selenium==3.141.0
six==1.12.0
torch==1.2.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
